+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [5.8.1 (unreleased)](https://github.com/kdeldycke/workflows/compare/v5.8.0...main)
+
+> [!IMPORTANT]
+> This version is not released yet and is under active development.
+
 ## [5.8.0 (2026-02-11)](https://github.com/kdeldycke/workflows/compare/v5.7.2...v5.8.0)
 
 - Fix stale checkout in `bump-versions` causing merge conflicts after releases.